u272

+++ hotspot/src/share/vm/jfr/writers/jfrStreamWriterHost.hpp
@@ -33,9 +33,9 @@ class StreamWriterHost : public MemoryWriterHost<Adapter, AP> {
  public:
   typedef typename Adapter::StorageType StorageType;
  private:
-  int64_t _stream_pos;
+  intptr_t _stream_pos;
   fio_fd _fd;
-  int64_t current_stream_position() const;
+  intptr_t current_stream_position() const;
 
  protected:
   StreamWriterHost(StorageType* storage, Thread* thread);
@@ -47,8 +47,8 @@ class StreamWriterHost : public MemoryWriterHost<Adapter, AP> {
   bool has_valid_fd() const;
 
  public:
-  int64_t current_offset() const;
-  void seek(int64_t offset);
+  intptr_t current_offset() const;
+  void seek(intptr_t offset);
   void flush();
   void write_unbuffered(const void* src, size_t len);
   bool is_valid() const;
